huata/requirements.txt
# Huata Calendar Agent Dependencies
# LLM-powered natural language calendar management

# Core dependencies
anthropic==0.39.0
redis==5.0.1
python-dotenv==1.0.0

# Google Calendar API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.109.0

# OAuth token encryption
cryptography==41.0.7

# Async support
aioredis==2.0.1
asyncio==3.4.3

# Web server for API endpoints
fastapi==0.104.1
uvicorn==0.24.0.post1

# Date/time utilities
python-dateutil==2.8.2
pytz==2023.3.post1

# JSON handling
orjson==3.9.10

# Development & testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
